livekit-agent-Voice/pyproject.toml
[project]
name = "livekit-agent-voice"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.5.7",
    "langchain-text-splitters>=1.1.1",
    "livekit-agents[silero,turn-detector]~=1.3",
    "livekit-plugins-cartesia>=1.5.1",
    "livekit-plugins-deepgram>=1.5.1",
    "livekit-plugins-google>=1.5.1",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
]
